feat: refactor file & error handling
