+++ rust/minigrep/src/main.rs
@@ -1,10 +1,5 @@
-use std::{env, fs};
-
-struct Config {
-    query: String,
-    file_path: String,
-}
-
+use std::env;
+use minigrep::Config;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -16,26 +11,10 @@ fn main() {
 
     println!("Searching for {}", config.query);
     println!("In file {}", config.file_path);
-
-    run(config);
-}
-
-
-impl Config {
-    fn build(args: &[String]) -> Result<Config, &'static str> {
-        if args.len() < 3 {
-            return Err("Not enough arguments");
-        }
-        let query = args[1].clone();
-        let file_path = args[2].clone();
-
-        Ok(Config { query, file_path })
+    if let Err(e) = minigrep::run(config) {
+        println!("Application error: {}",e);
+        std::process::exit(1);
     }
 }
 
-fn run(config: Config) {
-    let contents = fs::read_to_string(config.file_path)
-        .expect("Should have been able to read the file");
 
-    println!("With text:\n{contents}");
-}
